requirements.txt
Django
pytest
pytest-django
psycopg2-binary
google-generativeai
numpy
python-dotenv
tiktoken
coverage